stabilize

+++ src/layers/doubleSymLayer.jl
@@ -188,14 +188,16 @@ end
 function JTmv(this::DoubleSymLayer{T}, Zin::Array{T}, dummy::Array{T},
                                             theta::Array{T}, Yin::Array{T}, tmp) where {T<:Number}
 
-    nex::Int  = div(length(Yin),nFeatIn(this))
+    nex  = div(length(Yin),nFeatIn(this))
     Z         = reshape(Zin, :, nex)
     Yt        = reshape(tmp[2]::Array{T,2},:,nex)
     Y         = reshape(Yin,:,nex)
     th1, th2, th3, th4  = splitWeights(this,theta)
     Kop       = getOp(this.K,th1)
     A::Array{T,2}, dA::Array{T,2}    = this.activation(Yt,true)
+
     dth3      = vec(sum(this.Bout'*Z,2))
+
     dAZ1::Array{T,2}       = dA.*(Kop*Z)
     dth2      = vec(sum(this.Bin'*dAZ1,2))
     dth4, dAZ2::Vector{T}  = JTmv(this.nLayer,dAZ1,zeros(T,0),th4,Kop*Y,tmp[1])
